Created inventory directory and associated files
Created the basis of an inventory system, using inventory.java and
invMethods.java, where inventory.java inherits from invMethods.java.
My plan is to build a simple point of sale system using ArrayList to
work on the foundations of Java, and then further elaborate on it in
later projects.

To Do:
 - Edit inventory.java to throw error when non-integer input is provided
 - Build switch statements and determine inventory functions
 - Use ArrayList to track items currently in inventory