Implemented inventory view into purchases
The inventory is now viewable in the inventory.java file, where it was
implemented using methods in the invMethods.java file. Each object
contained within the pre-generated inventory is referenced by a specific
ID number using a HashMap. Next, I will begin working on referencing
and adding the inventory system to the sell portion of inventory.java,
and then I'll bounce back and begin work on the rest of the purchase
functions.