Edited inventory.java to include sell functions
I began creating the sell method, as well as doing some alterations
in the invMethods.java file to the purchase methods, where I added some
necessary options to the function. My plan is to implement the sell
methods next, and then work on building the cart function for the
purchase menu.