Implemented sell method and printHashMap in Help
The sell method was implemented in invMethods.java, where I asked for an
input and added it to the hashmap. This will reset when the program is
reset, which is something I need to look into at a later date. I also
implemented the Help portion to assist me in troubleshooting the
inventory program, where I can make methods that help me debug and other
sorts of things.

No issues fixed this commit, just features added.