Built "add to cart" function in inventory
I added the cart function in inventory by creating another hashmap that
holds the values for the cart. I then built a method that allows you to
give an input, and then adds that input to the hashmap, which then
prints the item you added as well as the corresponding hashmap value.

I need to implement some features to this that detect if an item is in
the inventory to add, as well as being able to search for things similar
using the variable instead of the hashmap ID.

I will be implementing a few things soon, which are:
 -Something to reference the objects in the cart hashmap, to verify if
  a cart item exists in the inventory
 -A way to search for objects, as well as a way to search without the
  hashmap ID.