Added 'remove from cart' to inventory.java
I implemented the remove from cart feature, where I printed the existing
values in the cart, and then removed the item that was provided as
input. I did this by calling the hashmap values and removing the value
given as an input. The StackOverflow question I referenced
(https://stackoverflow.com/questions/9527565/how-to-remove-entry-from
-hashmap-by-value) stated that it would use 'iteration internally,'
and I don't really know what it means, so I'm hoping it's not bad.

Things to add next:
 -Implement checkout function
 -Implement feature to check if item is in inventory before adding to
  the cart