Started working on cart function (buy menu 4)
Started working on the cart function, using a given input to select a few options. It isn't working for some reason, but I have class in a few, so I will work on it later.

I need to build the method, as well as fixing the if/else statement to actually work.