Added cart method to invMethods
I implemented the cart method in a way which it is easy to debug. It takes all the values from storage and the cart, prints them, checks the confirmation, then uses a for loop to check if the item is in the inventory, and if it is, it is removed. It will iterate through the entire storage array and then remove everything that should be removed. I checked immediately after and it printed the correct hashmap with the object removed.

When I went to option 0 in the main menu and printed the hashmap, the full hashmap was printed, however. I'll have to fix this error before moving on  to other systems.