Removed sleep from files and started purchaseLog
Sleep was removed from files. In hindsight, it was kind of redundant to
make the user wait for terminal outputs. The purchaseLog was started,
which will add a new method to the end of every buy and sell and store
them in the log, which will also be outputted like a text file.