Finished purchaseLog feature and updated README.md
Finished implementing purchase log, so I'm officially done with this
step of the project. There are definitely a ton of errors I didn't
100% comb through, but I think the general functionality works fine.
Also updated README.md to account for updated project goals and format
of ideas.

Known errors:
 -storage hashmap values aren't edited once removed VIA cart hashmap
 -any item can be bought/sold
 -names are formatted as buy/sell instead of add/remove, it's an
  inventory system not a POS system

Will be one to two more commits for release stuff but this is the last
that will edit files in the repository.